Move LidInfo and related classes to separate .h and .cpp files.
